added a cofirm before delete on cart page

+++ public/assets/js/cart.js
@@ -134,6 +134,10 @@ setItemTotal();
 setResume();
 
 const deleteItem = (productId) => {
+  const confirmDelete = confirm("Seguro desea eliminar el articulo?");
+
+  if (!confirmDelete) return;
+
   fetch(`${HOST}/shop/cart`, {
     method: "DELETE",
     headers: {
